fix(qa-lab): satisfy transport response checks
